Adding custom hook for generating history charts and using it in homepage
